requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
sqlalchemy>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
